Extracted fakes to separate files to reuse in tests
