+++ TestePratico.Domain.Tests/Fakes/FakeEntity.cs
@@ -1,0 +1,12 @@
+using TestePratico.Domain.Entities;
+using TestePratico.Domain.Validation;
+
+namespace TestePratico.Domain.Tests.Fakes;
+
+public class FakeEntity : EntityBase<FakeEntity>
+{
+    public FakeEntity(bool valid) : base(new FakeEntityValidator(valid))
+    {
+    }
+}
+
