+++ TestePratico.Domain.Tests/Services/ServiceBase_Add_Should.cs
@@ -1,45 +1,19 @@
+using Moq;
 using TestePratico.Domain.Interfaces;
 using TestePratico.Domain.Services;
-using Moq;
-using Moq.Protected;
-using TestePratico.Domain.Entities;
+using TestePratico.Domain.Tests.Fakes;
 using TestePratico.Domain.Validation;
-using TestePratico.Domain.Interfaces.Validation;
 
 namespace TestePratico.Domain.Tests.Services;
 
 public class ServiceBase_Add_Should
 {
-    FakeEntity validEntity = GetFakeEntity(valid: true);
-    FakeEntity invalidEntity = GetFakeEntity(valid: false);
-
-    public class FakeEntity : EntityBase<FakeEntity>
-    {
-        public FakeEntity(Validator<FakeEntity> validator) : base(validator)
-        {
-        }
-    }
-
-    public class FakeEntityValidator : Validator<FakeEntity>
-    {
-        private readonly bool valid;
-
-        public FakeEntityValidator(bool valid)
-        {
-            this.valid = valid;
-        }
-
-        public override ValidationResult Validate(FakeEntity entity)
-        {
-            return GetValidationResult(valid);
-        }
-    }
-
     [Fact]
     public void Return_Valid_ValidationResult_When_Entity_Is_Valid()
     {
         var sbugRepo = new Mock<IRepositoryBase<FakeEntity>>();
         var service = new ServiceBase<FakeEntity>(sbugRepo.Object);
+        var validEntity = new FakeEntity(valid: true);
 
         var result = service.Add(validEntity);
 
@@ -51,6 +25,7 @@ public class ServiceBase_Add_Should
     {
         var sbugRepo = new Mock<IRepositoryBase<FakeEntity>>();
         var service = new ServiceBase<FakeEntity>(sbugRepo.Object);
+        var invalidEntity = new FakeEntity(valid: false);
 
         var result = service.Add(invalidEntity);
 
@@ -63,6 +38,7 @@ public class ServiceBase_Add_Should
     {
         var mockRepo = new Mock<IRepositoryBase<FakeEntity>>();
         var service = new ServiceBase<FakeEntity>(mockRepo.Object);
+        var validEntity = new FakeEntity(valid: true);
 
         service.Add(validEntity);
 
@@ -76,25 +52,10 @@ public class ServiceBase_Add_Should
     {
         var mockRepo = new Mock<IRepositoryBase<FakeEntity>>();
         var service = new ServiceBase<FakeEntity>(mockRepo.Object);
+        var invalidEntity = new FakeEntity(valid: false);
 
         service.Add(invalidEntity);
 
         mockRepo.Verify(r => r.Add(It.IsAny<FakeEntity>()), Times.Never);
-    }
-
-
-    static FakeEntity GetFakeEntity(bool valid)
-    {
-        var validator = new FakeEntityValidator(valid);
-        return new FakeEntity(validator);
-    }
-
-    static ValidationResult GetValidationResult(bool valid)
-    {
-        var result = new ValidationResult();
-
-        if (!valid) result.Add("foo", "some error");
-
-        return result;
     }
 }
